Introduce kubernetes-node containers
This commit introduces containers for the Kubernetes node
services.  The node container contains the requistite package
to install the node service.  There is a container for each service
(kubelet and kube-proxy) that contain additional packages required to run
each service, as well as an atomic run label and an entrypoint for convinience
(these containers are based off of the node container).

Additionally, the non-base containers (i.e. all containers besides the
base master and node containers) have a default endpoint which pulls
configuration data from `/etc/kubernetes`.  This has the advantage of
allowing existing configuration to be reused (by mounting
`/etc/kubernetes:/etc/kubernetes:ro`), and of providing sensible
defaults when no options are given.

Signed-off-by: Jason Brooks <jbrooks@redhat.com>