Smoothened the branch to work follow the upstream/master.
Due to the fact the the CentOS Community Container pipeline was dependant
on the master branch of the fork of upstream, i had to make my changes there.
But since the same have not yet been pulled upstream, i have reverted those
changes in this branch so that it looks exactly like upstream/master.