Added functional elastic search container with v2.4 on openshift.
Currently only 2.4 is functional on openshift due to issues with
sysctl. I have created the 5.x container as well but its not
currently operational on openshift, but it works if sysctl is
configured with vm.max_map_count=262144. So more work will need
to be done once kubernetes fixes sysctl issue.